Fix jstype = JS_STRING field option for speed optimized code
There was a bug in internal-binary-read.ts

When the field option jstype = JS_STRING was set, the generated method `internalBinaryRead` would still read the field as a bigint.

This problem only surfaces if the plugin parameter long_type_string is *not* set and code is optimized for speed (instead of for code size).