Use details from service error as RpcError message
It seems that the `message` field of grpc-js errors contains the status code concatenated with the message. An error with invalid argument for example contains the following in `message`: `3 INVALID_ARGUMENT: something`.

It seems that the `details` field in the service error does not include that prefix, so I think that field should be used when constructing the RpcError.