Remove Angular support
I've personally had good experiences with Angular in the past, but it is time to remove support from this project. Weekly downloads on npmjs.com suggest that only 0.1% of protobuf-ts users use it, but it is taking much more than 0.1% percent of the maintance effort. Apologies to the few remaining users.