Simplify protobuf-ts.proto options
We define a few custom options in protobuf-ts.proto that affect code generation.

They are read from descriptors given to the plugin in interpreter.ts. Two interfaces and two enums are defined to match the types in protobuf-ts.proto.

This removes the 4 definitions with types generated from protobuf-ts.proto,

This also removes the OptionResolver class, moving the methods into InternalOptions.